Create markdown docs (#131)
As we slowly move towards the `1.0.0` goals, it will be very important to have a docs system with versioning. Our current docs site doesn't have this right now. While the long term goal is to make use of a good static site generator out there, it's not in my priority to look into it right now so a simple markdown docs in this repository will do.

With a simple markdown docs, we get versioning for free as well.

I've also took the time to update the docs with the latest Ember syntax as well as having guides for using the Firebase Emulator.